+++ curve25519-dalek/src/field.rs
@@ -25,8 +25,6 @@
 
 #[allow(unused_qualifications)]
 
-use core::cmp::{Eq, PartialEq};
-
 use cfg_if::cfg_if;
 
 use subtle::Choice;
@@ -322,7 +320,6 @@ extern crate rand;
 #[cfg(test)]
 mod test {
     use crate::field::*;
-    use subtle::ConditionallyNegatable;
 
     /// Random element a of GF(2^255-19), from Sage
     /// a = 1070314506888354081329385823235218444233221\